로그인 리다이렉트 구현

+++ src/main/java/com/example/dy/Service/ArticleService.java
@@ -2,14 +2,18 @@ package com.example.dy.Service;
 
 
 import com.example.dy.Domain.Article;
+import com.example.dy.Domain.User;
+import com.example.dy.Domain.constant.Role;
 import com.example.dy.Domain.constant.SearchType;
 import com.example.dy.Dto.ArticleDto;
 import com.example.dy.Repository.ArticleRepository;
-import com.example.dy.Repository.CommentRepository;
+import com.example.dy.Repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 
@@ -22,13 +26,14 @@ import javax.transaction.Transactional;
 @Service // 서비스 선언!(서비스 객체를 스프링부트에 생성)
 public class ArticleService {
 
-//    @Autowired // 필드 주입 방식 DI
-//    private ArticleRepository articleRepository;
-
 
     private final ArticleRepository articleRepository;
-    public ArticleService(ArticleRepository articleRepository) {
+    private final UserRepository userRepository; // UserRepository 주입
+    private final UserService userService;
+    public ArticleService(ArticleRepository articleRepository,UserRepository userRepository,UserService userService) {
         this.articleRepository = articleRepository;
+        this.userRepository = userRepository;
+        this.userService = userService;
     }
 
 
@@ -66,67 +71,76 @@ public class ArticleService {
     }
 
     //    2. 생성
-
     public ArticleDto create(ArticleDto dto) {
-
         log.info("생성 변환 이전(Dto) 1 : {}",dto);
+        // 현재 로그인한 사용자의 인증 정보 가져오기
+        User currentUser = userService.getCurrentUser(); // 현재 로그인한 사용자 정보 가져오기
+
         // 1.Dto를 사용하여 엔티티 생성
-        Article article = new Article(null, dto.getTitle(), dto.getContent());
+        Article article = new Article(null, dto.getTitle(), dto.getContent(), currentUser);
+
         log.info("생성 변환 이전(Entity) 2 : {}",article);
+
         // 2.엔티티 저장
-        Article createArticle = articleRepository.save(article);
-        // 3.엔티티를 DTO로 변환하여 반환
-        return ArticleDto.fromEntity(createArticle);
+        Article createdArticle = articleRepository.save(article);
+
+        // 3.생성된 Article 객체를 DTO로 변환하여 반환
+        return ArticleDto.fromEntity(createdArticle);
     }
 
 
 
 
-
-    //    3. 수정
+    // 3. 수정
     @Transactional
     public ArticleDto update(Long id, ArticleDto dto) {
-        // 1.ID로 기존 엔티티 조회
-        //   예외가 있으면 다음으로 안넘어감.
-        Article article = articleRepository.findById(id)
+
+            // 1.현재 로그인한 사용자의 인증 정보 가져오기
+        User currentUser = userService.getCurrentUser();
+            // 2.ID로 기존 엔티티 조회
+        Article target = articleRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Article not found with id: " + id));
-        log.info("수정 이전(Entity) 0 : {}",article);
 
-        // 2.엔티티 수정
-        Article articleToUpdate = new Article(dto.getId(), dto.getTitle(), dto.getContent());
-        log.info("수정 변환 이전(Entity) 1 : {}",articleToUpdate);
+            // 3.엔티티 수정
+        Article articleToUpdate = new Article(dto.getId(), dto.getTitle(), dto.getContent(), null);
 
-        article.patch(articleToUpdate);
+        // 게시글 작성자와 현재 로그인한 사용자 비교 하거나 Admin 아이디 인지 확인
+        if (currentUser.getRole() == Role.ADMIN || target.getUser().equals(currentUser)) {
+            target.patch(articleToUpdate);  // user는 변경하지 않음
 
-        log.info("수정 변환 이전(Entity) 2 : {}",article);
+            // 4.엔티티 저장
+            Article updatedArticle = articleRepository.save(target);
 
-
-        // 3.엔티티 저장
-        Article updatedArticle = articleRepository.save(article);
-
-        // 4.엔티티를 DTO로 변환하여 반환
-        return ArticleDto.fromEntity(updatedArticle);
+            // 5.엔티티를 DTO로 변환하여 반환
+            return ArticleDto.fromEntity(updatedArticle);
+        } else {
+            throw new IllegalStateException("수정 권한이 없습니다.");
+        }
     }
-
-
-
 
 
     //    4. 삭제
     public ArticleDto delete(Long id) {
 
-        // 1.대상 찾기 없으면 에러발생
+        // 1.현재 로그인한 사용자의 인증 정보 가져오기
+        User currentUser = userService.getCurrentUser(); // 현재 로그인한 사용자 정보 가져오기
+
+        // 2.대상 찾기 없으면 에러발생
         Article target = articleRepository.findById(id).
                 orElseThrow(() -> new EntityNotFoundException("Article not found with id: " + id));
+
+        // 3.댓글 작성자와 현재 로그인한 사용자 비교
+        if (!(currentUser.getRole() == Role.ADMIN) && !target.getUser().equals(currentUser)) {
+            throw new IllegalStateException("댓글 수정 실패! 자신이 작성한 댓글만 수정할 수 있습니다.");
+        }
         log.info("게시글 삭제 대상(Entity) 1: {}", target);
 
-        // 2. 대상을 삭제한다 !
+        // 4. 대상을 삭제한다 !
         articleRepository.delete(target);
 
-        // 3. 대상 변환 삭제 반환
+        // 5. 대상 변환 삭제 반환
         return ArticleDto.fromEntity(target);
     }
-
 
 }
