+++ src/main/java/com/example/dy/Service/CommentService.java
@@ -3,6 +3,8 @@ package com.example.dy.Service;
 
 import com.example.dy.Domain.Article;
 import com.example.dy.Domain.Comment;
+import com.example.dy.Domain.User;
+import com.example.dy.Domain.constant.Role;
 import com.example.dy.Dto.CommentDto;
 import com.example.dy.Repository.ArticleRepository;
 import com.example.dy.Repository.CommentRepository;
@@ -10,10 +12,13 @@ import com.example.dy.Repository.UserRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -31,11 +36,13 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final ArticleRepository articleRepository;
+    private final UserService userService;
 
     // 생성자 주입 방식 (권장)
-    public CommentService(CommentRepository commentRepository, ArticleRepository articleRepository) {
+    public CommentService(CommentRepository commentRepository, ArticleRepository articleRepository,UserService userService) {
         this.commentRepository = commentRepository;
         this.articleRepository = articleRepository;
+        this.userService = userService;
     }
 
 
@@ -73,12 +80,17 @@ public class CommentService {
         log.info("1. 게시글 Id : {}", articleId);
         log.info("2. 댓글 Dto : {}", dto);
 
+        log.info("생성 변환 이전(Dto) 1 : {}",dto);
+
+        // 현재 로그인한 사용자의 인증 정보 가져오기
+        User currentUser = userService.getCurrentUser(); // 현재 로그인한 사용자 정보 가져오기
+
         // 1.게시글 조회 및 예외 발생 (예외발생 되면 아래 실행X) -> comment 가 달린 위치 찾기 위해
         Article article = articleRepository.findById(articleId)
                 .orElseThrow(() -> new IllegalArgumentException("댓글 생성 실패! 대상 게시글이 없습니다."));
 
         // 2.댓글 엔티티 생성
-        Comment comment = Comment.createComment(dto, article); //  dto, article 두개다 던져줘서 댓글 엔티티 생성
+        Comment comment = Comment.createComment(dto, article,currentUser); //  dto, article 두개다 던져줘서 댓글 엔티티 생성
 
         // 3.댓글 엔티티를 DB로 저장
         Comment created = commentRepository.save(comment);
@@ -97,11 +109,21 @@ public class CommentService {
 
         log.info("1. 댓글 Dto : {}", dto);
 
+        // 현재 로그인한 사용자의 인증 정보 가져오기
+        User currentUser = userService.getCurrentUser(); // 현재 로그인한 사용자 정보 가져오기
+
         // 1.댓글 조회 및 예외 발생
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("댓글 수정 실패! 대상 댓글이 없습니다."));
 
         log.info("2. 댓글 엔티티 : {}",comment);
+
+        // 댓글 작성자와 현재 로그인한 사용자 비교
+        if (!(currentUser.getRole() == Role.ADMIN) && !comment.getUser().equals(currentUser)) {
+            throw new IllegalStateException("댓글 수정 실패! 자신이 작성한 댓글만 수정할 수 있습니다.");
+        }
+
+        log.info("서로 다른가? :{} {} {}",!comment.getUser().equals(currentUser),comment.getUser(),currentUser);
 
         // 2.댓글 수정
         comment.patch(dto);
@@ -116,9 +138,19 @@ public class CommentService {
 
     @Transactional
     public CommentDto delete(Long id) {
+
+        // 현재 로그인한 사용자의 인증 정보 가져오기
+        User currentUser = userService.getCurrentUser(); // 현재 로그인한 사용자 정보 가져오기
+
         // 댓글 조회(및 예외 발생)
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("댓글 삭제 실패! 대상이 없습니다."));
+
+        // 댓글 작성자와 현재 로그인한 사용자 비교
+        if (!(currentUser.getRole() == Role.ADMIN) && !comment.getUser().equals(currentUser)) {
+            throw new IllegalStateException("댓글 수정 실패! 자신이 작성한 댓글만 수정할 수 있습니다.");
+        }
+
         log.info("댓글 엔티티 =>{}",comment);
         // 댓글 삭제
         commentRepository.delete(comment);
